+++ sgl-kernel/tests/test_qserve_w4a8_per_group_gemm.py
@@ -1,3 +1,5 @@
+from test.utils import sym_quantize_tensor
+
 import pytest
 import torch
 from sgl_kernel import qserve_w4a8_per_group_gemm
@@ -117,13 +119,6 @@ def progressive_group_quantize_tensor(tensor, group_size):
         scale_i8.reshape(tensor.shape[0], -1).to(torch.int8),
         zero_i8.reshape(tensor.shape[0], -1).to(torch.int8),
     )
-
-
-# INT8 Quantization
-def sym_quantize_tensor(tensor):
-    tensor_scale = tensor.abs().max(dim=-1, keepdim=True)[0] / 127
-    tensor_q = torch.clamp(torch.round(tensor / tensor_scale), -128, 127).to(torch.int8)
-    return tensor_q, tensor_scale.to(torch.float16)
 
 
 def torch_w4a8_per_group_gemm(
