+++ test/registered/amd/perf/mi35x/test_grok1_int4_perf_mi35x.py
@@ -7,10 +7,9 @@ Registry: nightly-perf-8-gpu-mi35x-grok1-int4 suite
 
 import os
 import unittest
-from typing import List
+from test.registered.amd.perf.perf_utils import generate_simple_markdown_report
 
 from sglang.test.ci.ci_register import register_amd_ci
-from sglang.test.nightly_bench_utils import BenchmarkResult
 from sglang.test.nightly_utils import NightlyBenchmarkRunner
 from sglang.test.test_utils import DEFAULT_URL_FOR_TEST, _parse_int_list_env
 
@@ -18,27 +17,6 @@ from sglang.test.test_utils import DEFAULT_URL_FOR_TEST, _parse_int_list_env
 register_amd_ci(
     est_time=1500, suite="nightly-perf-8-gpu-mi35x-grok1-int4", nightly=True
 )
-
-
-def generate_simple_markdown_report(results: List[BenchmarkResult]) -> str:
-    """Generate a simplified markdown report without traces and cost columns."""
-    model_header = results[0].model_path
-    if results[0].run_name and results[0].run_name != "default":
-        model_header += f" ({results[0].run_name})"
-
-    gpu_config = os.getenv("GPU_CONFIG", "MI35x")
-    if gpu_config:
-        model_header += f" [{gpu_config}]"
-
-    summary = f"### {model_header}\n"
-    summary += "| batch size | input len | latency (s) | input throughput (tok/s) | output throughput (tok/s) | ITL (ms) |\n"
-    summary += "| ---------- | --------- | ----------- | ------------------------ | ------------------------- | -------- |\n"
-
-    for result in results:
-        itl = 1 / (result.output_throughput / result.batch_size) * 1000
-        summary += f"| {result.batch_size} | {result.input_len} | {result.latency:.2f} | {result.input_throughput:.2f} | {result.output_throughput:.2f} | {itl:.2f} |\n"
-
-    return summary
 
 
 # Model and tokenizer paths can be overridden via environment variables
@@ -108,7 +86,7 @@ class TestGrok1INT4PerfMI35x(unittest.TestCase):
 
             if results:
                 self.runner.full_report += (
-                    generate_simple_markdown_report(results) + "\n"
+                    generate_simple_markdown_report(results, "MI35x") + "\n"
                 )
 
             self.assertTrue(success, "Benchmark failed for Grok-1 INT4 on MI35x")
