+++ test/registered/amd/perf/test_grok1_fp8_perf.py
@@ -12,36 +12,14 @@ Example usage:
 
 import os
 import unittest
-from typing import List
+from test.registered.amd.perf.perf_utils import generate_simple_markdown_report
 
 from sglang.test.ci.ci_register import register_amd_ci
-from sglang.test.nightly_bench_utils import BenchmarkResult
 from sglang.test.nightly_utils import NightlyBenchmarkRunner
 from sglang.test.test_utils import DEFAULT_URL_FOR_TEST, _parse_int_list_env
 
 # Register for AMD CI - Grok-1 FP8 benchmark (~25 min)
 register_amd_ci(est_time=1500, suite="nightly-perf-8-gpu-grok1-fp8", nightly=True)
-
-
-def generate_simple_markdown_report(results: List[BenchmarkResult]) -> str:
-    """Generate a simplified markdown report without traces and cost columns."""
-    model_header = results[0].model_path
-    if results[0].run_name and results[0].run_name != "default":
-        model_header += f" ({results[0].run_name})"
-
-    gpu_config = os.getenv("GPU_CONFIG", "")
-    if gpu_config:
-        model_header += f" [{gpu_config}]"
-
-    summary = f"### {model_header}\n"
-    summary += "| batch size | input len | latency (s) | input throughput (tok/s) | output throughput (tok/s) | ITL (ms) |\n"
-    summary += "| ---------- | --------- | ----------- | ------------------------ | ------------------------- | -------- |\n"
-
-    for result in results:
-        itl = 1 / (result.output_throughput / result.batch_size) * 1000
-        summary += f"| {result.batch_size} | {result.input_len} | {result.latency:.2f} | {result.input_throughput:.2f} | {result.output_throughput:.2f} | {itl:.2f} |\n"
-
-    return summary
 
 
 # Model and tokenizer paths can be overridden via environment variables
