+++ test/registered/kernels/test_fp4_moe.py
@@ -3,15 +3,15 @@ from typing import Callable
 
 import pytest
 import torch
-from flashinfer import fp4_quantize, scaled_fp4_grouped_quantize
+from flashinfer import scaled_fp4_grouped_quantize
 from flashinfer.fused_moe import cutlass_fused_moe as flashinfer_cutlass_fused_moe
 from sgl_kernel import scaled_fp4_quant, silu_and_mul
-from torch.nn import functional as F
 
 from sglang.srt.layers.moe.cutlass_moe import cutlass_moe_fp4
 from sglang.srt.layers.moe.cutlass_moe_params import CutlassMoEParams, CutlassMoEType
 from sglang.srt.layers.moe.topk import TopKConfig, select_experts
 from sglang.test.ci.ci_register import register_cuda_ci
+from sglang.test.registered.kernels.moe.moe_utils import dequantize_nvfp4_to_dtype
 
 register_cuda_ci(est_time=300, suite="nightly-4-gpu-b200", nightly=True)
 
@@ -21,96 +21,8 @@ if torch.cuda.get_device_capability() < (10, 0):
         allow_module_level=True,
     )
 
-kE2M1ToFloat = torch.tensor(
-    [0.0, 0.5, 1.0, 1.5, 2.0, 3.0, 4.0, 6.0], dtype=torch.float32
-)
-
 FLOAT8_E4M3_MAX = 448.0
 FLOAT4_E2M1_MAX = 6.0
-
-
-def convert_swizzled_to_linear(a_sf_swizzled: torch.Tensor, m, k, block_size):
-    m_tiles = (m + 128 - 1) // 128
-    f = block_size * 4
-    k_tiles = (k + f - 1) // f
-    tmp = torch.reshape(a_sf_swizzled, (1, m_tiles, k_tiles, 32, 4, 4))
-    tmp = torch.permute(tmp, (0, 1, 4, 3, 2, 5))
-    out = tmp.reshape(m_tiles * 128, k_tiles * f // block_size)
-    return out[0:m, 0:k]
-
-
-def dequantize_nvfp4_to_dtype(
-    tensor_fp4, tensor_sf, global_scale, dtype, device, block_size=16
-):
-    """Dequantize the fp4 tensor back to high precision."""
-    # Two fp4 values are packed into one uint8.
-    assert tensor_fp4.dtype == torch.uint8
-    m, packed_k = tensor_fp4.shape
-    k = packed_k * 2
-    tensor_f32 = break_fp4_bytes(tensor_fp4, dtype)
-    tensor_f32 = tensor_f32.reshape(m, k // block_size, block_size)
-    tensor_sf = tensor_sf.view(torch.float8_e4m3fn)
-    tensor_sf = convert_swizzled_to_linear(tensor_sf, m, k, block_size)
-    tensor_sf_dtype = tensor_sf.to(torch.float32) / global_scale
-
-    # scale the tensor
-    out = (tensor_f32 * tensor_sf_dtype.unsqueeze(-1)).reshape(m, k)
-    return out.to(dtype=dtype)
-
-
-def break_fp4_bytes(a, dtype):
-    assert a.dtype == torch.uint8
-    m, n = a.shape
-
-    # Vectorized nibble processing
-    a_flat = a.flatten()
-    high = (a_flat & 0xF0) >> 4  # Upper nibbles
-    low = a_flat & 0x0F  # Lower nibbles
-
-    # Combine nibbles for batch processing
-    combined = torch.stack((low, high), dim=1).flatten()
-
-    # Vectorized sign and magnitude extraction
-    signs = (combined & 0x08).to(torch.bool)  # Sign bits
-    abs_vals = (combined & 0x07).to(torch.long)  # Magnitude indices
-
-    # Device-aware lookup and sign application
-    kE2M1 = kE2M1ToFloat.to(device=a.device)
-    values = kE2M1[abs_vals] * torch.where(signs, -1.0, 1.0)
-
-    # Reshape to final form
-    return values.reshape(m, n * 2).to(dtype=dtype)
-
-
-def compute_routing(router_logits: torch.Tensor, top_k: int):
-    routing_weights = torch.softmax(router_logits, dim=1, dtype=torch.float)
-    routing_weights, selected_experts = torch.topk(routing_weights, top_k, dim=-1)
-    routing_weights /= routing_weights.sum(dim=-1, keepdim=True)
-    routing_weights = routing_weights.float()
-    return routing_weights, selected_experts
-
-
-def prepare_inputs(
-    hidden_states: torch.Tensor,
-    router_logits: torch.Tensor,
-    num_experts: int,
-    topk: int,
-):
-    routing_weights, topk_idx = compute_routing(router_logits, topk)
-
-    masked_m = []
-    for i in range(num_experts):
-        mask = topk_idx.view(-1) == i
-        masked_m.append(mask.sum())
-
-    masked_m = torch.tensor(masked_m, dtype=torch.int32)
-    hidden_states_3d = torch.empty(
-        (num_experts, max(masked_m), hidden_states.shape[1]), dtype=hidden_states.dtype
-    )
-    for i in range(num_experts):
-        hidden_states_3d[i, : masked_m[i], :] = hidden_states[topk_idx.view(-1) == i]
-
-    return hidden_states_3d, masked_m, topk_idx, routing_weights
 
 
 MNK_FACTORS = [
@@ -144,38 +56,6 @@ def torch_moe(a, w1, w2, score, topk, expert_map):
             out[mask] = silu_and_mul(a[mask] @ w1[i].transpose(0, 1)) @ w2[i].transpose(
                 0, 1
             )
-    return (
-        out.view(B, -1, w2.shape[1]) * topk_weight.view(B, -1, 1).to(out.dtype)
-    ).sum(dim=1)
-
-
-def torch_moe_nvfp4(a, w1, w2, topk, topk_weight, topk_ids):
-    B, D = a.shape
-    a = a.view(B, -1, D).repeat(1, topk, 1).reshape(-1, D)
-    out = torch.zeros(B * topk, w2.shape[1], dtype=a.dtype, device=a.device)
-
-    topk_weight = topk_weight.view(-1)
-    topk_ids = topk_ids.view(-1)
-
-    for i in range(w1.shape[0]):
-        mask = topk_ids == i
-        if mask.sum():
-            m = w1[i].shape[0]
-            assert m % 2 == 0
-            # Note: w1 and w3 are swapped!
-            w3_expert, w1_expert = w1[i][m // 2 :, :], w1[i][: m // 2, :]
-            inter = F.silu(a[mask] @ w1_expert.t()) * (a[mask] @ w3_expert.t())
-            inter_gs = torch.tensor(1.0).cuda()
-            inter_q, inter_blockscale = fp4_quantize(inter, inter_gs)
-            inter = dequantize_nvfp4_to_dtype(
-                inter_q,
-                inter_blockscale,
-                inter_gs,
-                dtype=inter.dtype,
-                device=inter.device,
-                block_size=16,
-            ).cuda()
-            out[mask] = inter @ w2[i].transpose(0, 1)
     return (
         out.view(B, -1, w2.shape[1]) * topk_weight.view(B, -1, 1).to(out.dtype)
     ).sum(dim=1)
